requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
numpy
scikit-learn
joblib
